+++ src/main/org/hornetq/api/jms/management/DestinationControl.java
@@ -45,7 +45,7 @@ public interface DestinationControl
    /**
     * Returns the number of messages currently in this destination.
     */
-   int getMessageCount() throws Exception;
+   long getMessageCount() throws Exception;
 
    /**
     * Returns the number of messages that this queue is currently delivering to its consumers.
