+++ src/main/org/hornetq/api/jms/management/JMSQueueControl.java
@@ -103,7 +103,7 @@ public interface JMSQueueControl extends DestinationControl
     * Using {@code null} or an empty filter will count <em>all</em> messages from this queue.
     */
    @Operation(desc = "Returns the number of the messages in the queue matching the given filter", impact = MBeanOperationInfo.INFO)
-   int countMessages(@Parameter(name = "filter", desc = "A JMS message filter (can be empty)") String filter) throws Exception;
+   long countMessages(@Parameter(name = "filter", desc = "A JMS message filter (can be empty)") String filter) throws Exception;
 
    /**
     * Removes the message corresponding to the specified message ID.
