+++ src/main/org/hornetq/core/server/Queue.java
@@ -72,7 +72,7 @@ public interface Queue extends Bindable
 
    void deliverAsync();
 
-   int getMessageCount();
+   long getMessageCount();
 
    int getDeliveringCount();
